Migrate to the tm4c123x-hal crate
The `Peripherals` struct is split into `board::Board` -- this is
necessary because GPIO_PORTF is moved when it is `.split()`, and thus
the whole Peripherals struct is partially-moved.  I tried creating
another Peripherals using `.steal()`, but if the user wants to override
behavior for only a single LED pin, their call to`GpioExt::split()` will
reset the entire GpioF power domain and destroy the remainder of the
configuration.

This also adds `build.rs` akin to cortex-m to provide `memory.x` to
cortex-m-rt for users of this crate.